requirements.txt
Django==2.1.4
django-jenkins==0.110.0
djangorestframework==3.9.0
djangorestframework-simplejwt==3.3
pep8==1.7.1
Pillow==5.3.0
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pyflakes==2.0.0
PyJWT==1.7.1
pytz==2018.7
